+++ src/main.ts
@@ -1,2 +1,2 @@
 export { src, IConfig } from "./src";
-export { IVirtPathMapping } from "./resolver";
+export { IPathMapper } from "./resolver";